refactored mapper
